requirements.txt
smolagents[toolkit,litellm]
wikipedia-api
wikipedia
requests
python-dotenv
arize-phoenix
openinference-instrumentation-litellm

